promoted sequence_simple to sequence
